DNN Install Package
This is the first release of the WET framework DNN variant.  It has been
tested and is fully functional in IE 8 +, chrome, and firefox.  The
administration bar has been updated to reflect the new WET features as
well.

It is not WPSS ready yet based on comment locations, but it is fine for
intranets.